pyproject.toml
[project]
name = "rucode-competition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "autogluon>=1.1.1",
    "catboost>=1.1.1",
    "datasets>=2.14.4",
    "deep-translator>=1.11.4",
    "evaluate>=0.4.3",
    "ipywidgets",
    "jupyterlab",
    "lightgbm>=4.3.0",
    "matplotlib",
    "nlpaug>=1.1.11",
    "nltk>=3.9.1",
    "numpy",
    "open-clip-torch>=2.26.1",
    "pandarallel>=1.6.5",
    "pandas",
    "pillow>=10.4.0",
    "plotly>=5.24.1",
    "prophet>=1.1.5",
    "protobuf>=5.28.2",
    "sacremoses>=0.1.1",
    "scikit-learn>=1.4.0",
    "sentencepiece>=0.2.0",
    "skforecast>=0.13.0",
    "sktime>=0.33.1",
    "spacy>=3.7.5",
    "statsmodels>=0.14.3",
    "torch>=2.3.1",
    "torchvision>=0.18.1",
    "tqdm>=4.66.5",
    "transformers>=4.40.2",
    "tsfresh>=0.20.2",
    "wandb>=0.18.3",
]

[tool.uv]
dev-dependencies = [
    "black",
    "deptry",
    "flake8",
    "flake8-bugbear",
    "flake8-pyproject",
    "flake8-simplify",
    "isort",
    "mypy",
    "nbqa[toolchain]",
    "pandas-stubs",
    "pre-commit",
    "pylint",
    "ruff",
    "types-beautifulsoup4",
    "types-colorama",
    "types-decorator",
    "types-jsonschema",
    "types-openpyxl",
    "types-pillow",
    "types-protobuf",
    "types-psutil",
    "types-pygments",
    "types-pyyaml",
    "types-regex",
    "types-requests",
    "types-seaborn",
    "types-setuptools",
    "types-six",
    "types-tabulate",
    "types-tqdm",
    "typing-extensions",
]

[tool.black]
line-length = 90
target-version = ["py39"]
# skip-magic-trailing-comma = true

[tool.isort]
profile = "black"
line_length = 90
float_to_top = true
known_third_party = ["wandb"]

[tool.autoflake]
expand-star-imports = true
remove-all-unused-imports = true
ignore-init-module-imports = true
remove-duplicate-keys = true
remove-unused-variables = true
recursive = true
in-place = true

[tool.flake8]
max-line-length = 90
extend-ignore = ["E203"] # Whitespace before ':' for black compatability

[tool.pylint]
py-version = "3.9"
max-line-length = 90
disable = [
  "C0116", # documentation
  "C0114", # documentation
  "C0115", # documentation
]

[tool.mypy]
python_version = "3.9"
strict = false            # it's too strict to be useful
show_error_context = true
pretty = true
# ignore_missing_imports = true # ignore packages without type-checking

[tool.ruff]
target-version = "py39"
line-length = 90

lint.select = ["ALL"]
lint.ignore = [
  "ANN101",  # annotationg self in class methods, will be deprecated
  "ANN102",  # annotationg cls in class methods, will be deprecated
  "COM812",  # missing-trailing-comma
  "D",       # documentation
  "FA100",   # future-annotations
  "NPY002",  # numpy random stuff
  "PLR0913", # too many init arguments
  "RUF012",  # mutable class attributes type checking
  "T201",    # print statements
]
